+++ src/pages/ManageStoreProduct/index.js
@@ -63,23 +63,36 @@ const ManageStoreProduct = () => {
     })
   }
   const deleteAllFunction = async (selected) => {
-    Swal.showLoading();
-    const listPromise = [];
-    selected.map((product) => {
-      listPromise.push(
-        new Promise((resolve) => {
-          dispatch(doDeleteProduct({
-            id: product
-            })).then(() => {
-                resolve();
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.showLoading();
+        const listPromise = [];
+        selected.map((product) => {
+          listPromise.push(
+            new Promise((resolve) => {
+              dispatch(doDeleteProduct({
+                id: product
+                })).then(() => {
+                    resolve();
+                })
             })
+          )
         })
-      )
-    })
-    Promise.all(listPromise).then(() => {
-      Swal.close();
-      returnTable();
-    })
+        Promise.all(listPromise).then(() => {
+          Swal.close();
+          returnTable();
+        })
+      }
+  })
+   
   }
   const returnTable = async () => {
     await dispatch(doGetListProductsOfStores({
